requirements.txt
wheel
pyaes
telethon
aiohttp
python-dotenv
pymongo
requests
